Crate futures_become

Tail-call elimination for futures.

Dispatching a long chain of futures built with and_then may result in a linear cost. One could avoid this using e.g. loop_fn, but this is hard to compose due to shared state.

This package provides an alternative interface through a Become construct, which acts as a kind of "tail-call elimination" at the futures level. Turns out it can be done without modifying the existing Future trait, but at the cost of changing Error type of each future to a wrapped one using Become. Then, returning Err(Become::Become(next)) can be used to indicate that the current future should be discarded and replaced with next. This of course means that we have to box each future, as the type of the current future could be wildly different from that of the next future.

extern crate futures;
extern crate futures_become;

use futures::{future, Async, Future};
use futures_become::{becomer, becomes, err};

fn main() {
    let mut start = true;
    let mut root = becomer(future::poll_fn(|| {
        if start {
            start = false;
            Ok(Async::Ready(true))
        } else {
            unreachable!()
        }
    }).and_then(|b| {
        assert!(b);
        becomes(future::ok(":3").and_then(|s| {
            if s == ":3" {
                future::ok(())
            } else {
                err(())
            }
        }))
    }));
    match root.poll() {
        Ok(Async::Ready(())) => (),
        _ => unreachable!(),
    }
}

Modules

unsync
Analog of the outer module, but for non-Send futures.

Structs

Becomer
Dispatcher that can perform dynamic tail-call elimination.

Traits

Becoming
Controls the type of trait object used to dispatch tail calls.

Functions

becomer
Create a new future that has the ability to “run” a subfuture that supports tail-call elimination. This introduces a Becomer context.

becomes
A Future that simply becomes its contained value. This only makes sense when run under a Becomer context.

err
A Future that simply raises an error. Replaces futures::future::err in Becomer contexts.

Type Definitions

BoxBecomer
An alias for a Becomer that is boxed and Send.

BoxFuture
An alias for a boxed Future that is Send.